chore(sales-context): adicionar verificações de nulidade e criar commands, events e handlers - Adicionar verificações de nulidade nos métodos GetProducts, GetProductsByCategoryName, GetProduct, AddProduct, UpdateProduct, DeleteProduct e GetCategories para evitar NullReferenceException. - Cria novos commands para operações de produto e categoria. - Cria novos events para operações de produto e categoria. - Cria novos handlers para processar commands e events de produto e categoria.

+++ src/Ambev.DeveloperEvaluation.ORM/Repositories/Catalog/ProductRepository.cs
@@ -23,7 +23,7 @@ namespace Ambev.DeveloperEvaluation.ORM.Repositories.Catalog
                 .ToListAsync();
         }
 
-        public async Task<Product> GetById(Guid id)
+        public async Task<Product?> GetById(Guid id)
         {
             var product = await _context.Products
                 .Include(p => p.Category)
@@ -65,7 +65,8 @@ namespace Ambev.DeveloperEvaluation.ORM.Repositories.Catalog
 
         public void DeleteProduct(Guid id)
         {
-            var product = _context.Products.FirstOrDefault(p => p.Id == id);            
+            var product = _context.Products.FirstOrDefault(p => p.Id == id);
+            if (product == null) throw new ArgumentNullException(nameof(product));
             _context.Products.Remove(product);            
             
         }
